Introduce git alias to unset local hooksPath
The hooksPath in the git user config already points to a non-existing
path so hooks in $REPO/.git/hooks are not executed. This worked well to
prevent tools like husky running hooks on my machine. But it seems like
the latest version of husky is defining hooksPath in the repo git config
which has precedence over the user config. This alias will remove the
local hooksPath configuration everytime git is running.